requirements.txt
asgiref==3.7.2
attrs==23.1.0
backports.zoneinfo==0.2.1
Django==4.2.7
django-filter==23.4
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
importlib-metadata==6.8.0
importlib-resources==6.1.1
inflection==0.5.1
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
Markdown==3.5.1
packaging==23.2
Pillow==10.1.0
pkgutil_resolve_name==1.3.10
PyJWT==2.8.0
python-decouple==3.8
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.32.0
rpds-py==0.13.2
sqlparse==0.4.4
swagger-spec-validator==3.0.3
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
zipp==3.17.0
